Implement proper "initialValue" handling
The way this works now is that observables don't store data.
They store procs that will emit data to observers.

"Cold" observables (and some subject types such as Behavior and Replaysubjects) are unique in the sense
that they want to emit those values immediately after an observer subscribes.

Therefore observables now have a field that checks if an Observable has initialValues.
If it does, then the initialHandler is called with the observable that was passed to subscribe.
That initialHandler should then emit all the initial data into the subscriber.

Given that operators exist, which create more observables and those observables may not themselves hold data,
these need to perform some special juice.
They need to implement their own "initialValues" handler which does nothing but fetch the data from their parent observable and forward it (potentially mutated) to the observer they received.

That way, after subscribing, the observer gets sent all the events from the observable it subscribes to,
that observable receives all the events from the parent it subscribes to (through a one-off "forwarding" observer) and so on and so forth
until the "root" initialHandler has been hit.